+++ attributes/default.rb
@@ -1,5 +1,5 @@
 default['selenium']['url'] =
-  'https://selenium-release.storage.googleapis.com/2.53/selenium-server-standalone-2.53.1.jar'
+  'https://selenium-release.storage.googleapis.com/3.0/selenium-server-standalone-3.0.1.jar'
 
 default['selenium']['windows']['home'] = "#{ENV['SYSTEMDRIVE']}/selenium"
 default['selenium']['windows']['java'] = "#{ENV['SYSTEMDRIVE']}\\java\\bin\\java.exe"
@@ -7,3 +7,11 @@ default['selenium']['windows']['java'] = "#{ENV['SYSTEMDRIVE']}\\java\\bin\\java
 # used by both macosx and linux platforms
 default['selenium']['unix']['home'] = '/opt/selenium'
 default['selenium']['unix']['java'] = '/usr/bin/java'
+
+if platform?('windows')
+  default['selenium']['home'] = node['selenium']['windows']['home']
+  default['selenium']['java'] = node['selenium']['windows']['java']
+else
+  default['selenium']['home'] = node['selenium']['unix']['home']
+  default['selenium']['java'] = node['selenium']['unix']['java']
+end