+++ providers/hub.rb
@@ -13,8 +13,10 @@ def config
       resource: new_resource
     )
     notifies :restart, "service[#{new_resource.servicename}]", :delayed unless platform_family?('windows', 'mac_os_x')
-    notifies :run, "execute[reload #{selenium_mac_domain(new_resource.servicename)}]",
-             :delayed if platform_family?('mac_os_x')
+    if platform_family?('mac_os_x')
+      notifies :run, "execute[reload #{selenium_mac_domain(new_resource.servicename)}]",
+               :delayed
+    end
   end
   config_file
 end
@@ -28,9 +30,11 @@ end
 
 action :install do
   converge_by("Install Hub Service: #{new_resource.servicename}") do
-    recipe_eval do
-      run_context.include_recipe 'selenium::default'
-    end unless run_context.loaded_recipe? 'selenium::default'
+    unless run_context.loaded_recipe? 'selenium::default'
+      recipe_eval do
+        run_context.include_recipe 'selenium::default'
+      end
+    end
 
     case node['platform']
     when 'windows'
@@ -38,7 +42,7 @@ action :install do
       selenium_windows_firewall(new_resource.servicename, new_resource.port)
     when 'mac_os_x'
       plist = "/Library/LaunchDaemons/#{selenium_mac_domain(new_resource.servicename)}.plist"
-      selenium_mac_service(selenium_mac_domain(new_resource.servicename), selenium_java_exec, args, plist, nil)
+      selenium_mac_service(new_resource, selenium_java_exec, args, plist, nil)
     else
       selenium_linux_service(new_resource.servicename, selenium_java_exec, args, new_resource.port, nil)
     end