+++ providers/node.rb
@@ -4,19 +4,28 @@ def whyrun_supported?
   true
 end
 
+def use_selenium2_syntax?
+  selenium_version = node['selenium']['url'][/standalone-(.*)\.jar/, 1]
+  Gem::Version.new(selenium_version) < Gem::Version.new('3')
+end
+
 def config
   config_file = "#{selenium_home}/config/#{new_resource.servicename}.json"
   template config_file do
     source 'node_config.erb'
     cookbook 'selenium'
     variables(
-      resource: new_resource
+      resource: new_resource,
+      use_selenium2_syntax: use_selenium2_syntax?
     )
-    notifies :request, "windows_reboot[Reboot to start #{new_resource.servicename}]",
-             :delayed if platform_family?('windows')
+    if platform_family?('windows')
+      notifies :request_reboot, "reboot[Reboot to start #{new_resource.servicename}]", :delayed
+    end
     notifies :restart, "service[#{new_resource.servicename}]", :delayed unless platform_family?('windows', 'mac_os_x')
-    notifies :run, "execute[reload #{selenium_mac_domain(new_resource.servicename)}]",
-             :immediately if platform_family?('mac_os_x')
+    if platform_family?('mac_os_x')
+      notifies :run, "execute[reload #{selenium_mac_domain(new_resource.servicename)}]",
+               :immediately
+    end
   end
   config_file
 end
@@ -25,35 +34,28 @@ def args
   args = []
   args << new_resource.jvm_args unless new_resource.jvm_args.nil?
   args << %W(-jar "#{selenium_jar_link}" -role node -nodeConfig "#{config}")
-
-  new_resource.additional_args.each do |arg|
-    args << arg
-  end
-
   args.flatten!
 end
 
 action :install do
   converge_by("Install Node Service: #{new_resource.servicename}") do
-    recipe_eval do
-      run_context.include_recipe 'selenium::default'
-    end unless run_context.loaded_recipe? 'selenium::default'
+    unless run_context.loaded_recipe? 'selenium::default'
+      recipe_eval do
+        run_context.include_recipe 'selenium::default'
+      end
+    end
 
     case node['platform']
     when 'windows'
-      recipe_eval do
-        run_context.include_recipe 'windows::reboot_handler'
-      end unless run_context.loaded_recipe? 'windows::reboot_handler'
-
       selenium_windows_gui_service(new_resource.servicename, selenium_java_exec, args, new_resource.username)
-      selenium_autologon(new_resource.username, new_resource.password, new_resource.domain)
+      selenium_autologon(new_resource.username, new_resource.password)
 
       selenium_windows_firewall(new_resource.servicename, new_resource.port)
 
-      windows_reboot "Reboot to start #{new_resource.servicename}" do
-        reason "Reboot to start #{new_resource.servicename}"
-        timeout node['windows']['reboot_timeout']
+      reboot "Reboot to start #{new_resource.servicename}" do
         action :nothing
+        reason 'Need to reboot when the run completes successfully.'
+        delay_mins 1
       end
     when 'mac_os_x'
       plist = if new_resource.username && new_resource.password
@@ -63,7 +65,8 @@ action :install do
               end
 
       selenium_mac_service(
-        selenium_mac_domain(new_resource.servicename), selenium_java_exec, args, plist, new_resource.username
+        selenium_mac_domain(new_resource.servicename), selenium_java_exec, args, plist,
+        new_resource.username, new_resource.log
       )
       selenium_autologon(new_resource.username, new_resource.password)