+++ spec/unit/hub_spec.rb
@@ -11,9 +11,9 @@ describe 'selenium_test::hub' do
     let(:chef_run) do
       ChefSpec::SoloRunner.new(platform: 'windows', version: '2008R2', step_into: ['selenium_hub']) do |node|
         ENV['SYSTEMDRIVE'] = 'C:'
-        node.set['selenium']['url'] =
-          'https://selenium-release.storage.googleapis.com/2.45/selenium-server-standalone-2.45.0.jar'
-        node.set['java']['windows']['url'] = 'http://ignore/jdk-windows-64x.tar.gz'
+        node.override['selenium']['url'] =
+          'https://selenium-release.storage.googleapis.com/3.0/selenium-server-standalone-3.0.1.jar'
+        node.override['java']['windows']['url'] = 'http://ignore/jdk-windows-64x.tar.gz'
         stub_command('netsh advfirewall firewall show rule name="selenium_hub" > nul').and_return(false)
       end.converge(described_recipe)
     end
@@ -35,10 +35,7 @@ describe 'selenium_test::hub' do
         args: '-jar """C:/selenium/server/selenium-server-standalone.jar"""'\
           ' -role hub -hubConfig """C:/selenium/config/selenium_hub.json"""',
         params: {
-          AppDirectory: 'C:/selenium',
-          AppStdout: 'C:/selenium/log/selenium_hub.log',
-          AppStderr: 'C:/selenium/log/selenium_hub.log',
-          AppRotateFiles: 1
+          AppDirectory: 'C:/selenium'
         }
       )
     end
@@ -48,15 +45,15 @@ describe 'selenium_test::hub' do
     end
 
     it 'reboots windows server' do
-      expect(chef_run).to_not request_windows_reboot('Reboot to start selenium_hub')
+      expect(chef_run).to_not request_reboot('Reboot to start selenium_hub')
     end
   end
 
   context 'linux' do
     let(:chef_run) do
       ChefSpec::SoloRunner.new(platform: 'centos', version: '7.0', step_into: ['selenium_hub']) do |node|
-        node.set['selenium']['url'] =
-          'https://selenium-release.storage.googleapis.com/2.45/selenium-server-standalone-2.45.0.jar'
+        node.override['selenium']['url'] =
+          'https://selenium-release.storage.googleapis.com/3.0/selenium-server-standalone-3.0.1.jar'
         allow_any_instance_of(Chef::Provider).to receive(:selenium_systype).and_return('systemd')
       end.converge(described_recipe)
     end
@@ -101,8 +98,9 @@ describe 'selenium_test::hub' do
   context 'mac_os_x' do
     let(:chef_run) do
       ChefSpec::SoloRunner.new(platform: 'mac_os_x', version: '10.10', step_into: ['selenium_hub']) do |node|
-        node.set['selenium']['url'] =
-          'https://selenium-release.storage.googleapis.com/2.45/selenium-server-standalone-2.45.0.jar'
+        node.override['selenium']['url'] =
+          'https://selenium-release.storage.googleapis.com/3.0/selenium-server-standalone-3.0.1.jar'
+        node.override['selenium']['hub']['log'] = '/var/log/selenium/org.seleniumhq.selenium_hub.log'
         allow_any_instance_of(Chef::Recipe).to receive(:java_version_on_macosx?).and_return(false)
       end.converge(described_recipe)
     end