+++ test/fixtures/cookbooks/selenium_test/metadata.rb
@@ -8,12 +8,9 @@ version '0.1.0'
 
 depends 'selenium'
 
-depends 'java_se', '~> 8.0'
-depends 'mozilla_firefox', '~> 1.0'
-
-depends 'xvfb', '~> 2.0'
-depends 'windows_screenresolution', '~> 1.0'
-
-depends 'apt'
-depends 'yum'
 depends 'dmg'
+depends 'geckodriver'
+depends 'java_se', '~> 8.0'
+depends 'mozilla_firefox'
+depends 'windows_screenresolution'
+depends 'xvfb'